Refactoring ContrabandHitRateHandler
